requirements.txt
# ==========================================
# Agente IIBB — Dependencias del Proyecto
# ==========================================

# LLM & Agente
openai>=1.30.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-google-genai>=1.0.0

# SDKs Nativos
google-generativeai>=0.5.0

# Vector Store (RAG local)
chromadb>=0.5.0

# Embeddings locales (alternativa sin costo a OpenAI embeddings)
sentence-transformers>=3.0.0

# PDF Parsing
pdfplumber>=0.11.0
pypdf>=4.0.0

# Terminal UI & Web App
rich>=13.7.0
streamlit>=1.35.0

# Procesamiento de Datos & Reportes
pandas>=2.0.0
xlsxwriter>=3.1.0
python-docx>=1.1.0
sqlalchemy>=2.0.0

# Utilidades
python-dotenv>=1.0.0
pydantic>=2.0.0
